refactor(mongo): extract date extract translation maps
