+++ src/lib/translators/mongo.ts
@@ -8,6 +8,7 @@ import { $$, combinations, escapeForUseInRegExp } from '@/lib/helpers';
 import { OutputStep, StepMatcher } from '@/lib/matcher';
 import * as S from '@/lib/steps';
 import { BaseTranslator, ValidationError } from '@/lib/translators/base';
+import { ADVANCED_DATE_EXTRACT_MAP, DATE_EXTRACT_MAP } from '@/lib/translators/mongo_dates';
 import { VariableDelimiters } from '@/lib/variables';
 
 type PropMap<T> = { [prop: string]: T };
@@ -23,32 +24,12 @@ type ComboOperator = 'and' | 'or';
 
 export type DateGranularity = 'day' | 'month' | 'year';
 
-type BasicDateOperationMap = {
-  [OP in S.BasicDatePart]: string;
-};
-
 type DurationMultiplierMap = {
   [D in S.ComputeDurationStep['durationIn']]: number;
 };
 
 type FilterComboAndMongo = {
   $and: MongoStep[];
-};
-
-const DATE_EXTRACT_MAP: BasicDateOperationMap = {
-  year: '$year',
-  month: '$month',
-  day: '$dayOfMonth',
-  week: '$week',
-  dayOfYear: '$dayOfYear',
-  dayOfWeek: '$dayOfWeek',
-  isoYear: '$isoWeekYear',
-  isoWeek: '$isoWeek',
-  isoDayOfWeek: '$isoDayOfWeek',
-  hour: '$hour',
-  minutes: '$minute',
-  seconds: '$second',
-  milliseconds: '$millisecond',
 };
 
 // A mapping of multiplier to apply to convert milliseconds in days, hours, minutes or seconds
@@ -647,179 +628,9 @@ function transformDateExtract(step: Readonly<S.DateExtractStep>): MongoStep {
 
   for (let i = 0; i < dateInfo.length; i++) {
     const d = dateInfo[i];
-    if (d === 'quarter') {
-      addFields[newColumns[i]] = {
-        $switch: {
-          branches: [
-            { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 1] }, then: 1 },
-            { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 2] }, then: 2 },
-            { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 3] }, then: 3 },
-          ],
-          default: 4,
-        },
-      };
-    } else if (d === 'firstDayOfYear') {
-      addFields[newColumns[i]] = {
-        $dateFromParts: { year: { $year: $$(step.column) }, month: 1, day: 1 },
-      };
-    } else if (d === 'firstDayOfMonth') {
-      addFields[newColumns[i]] = {
-        $dateFromParts: {
-          year: { $year: $$(step.column) },
-          month: { $month: $$(step.column) },
-          day: 1,
-        },
-      };
-    } else if (d === 'firstDayOfWeek') {
-      addFields[newColumns[i]] = {
-        // We subtract to the target date a number of days corresponding to (dayOfWeek - 1)
-        $subtract: [
-          $$(step.column),
-          {
-            $multiply: [{ $subtract: [{ $dayOfWeek: $$(step.column) }, 1] }, 24 * 60 * 60 * 1000],
-          },
-        ],
-      };
-    } else if (d === 'firstDayOfQuarter') {
-      addFields[newColumns[i]] = {
-        $dateFromParts: {
-          year: { $year: $$(step.column) },
-          month: {
-            $switch: {
-              branches: [
-                { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 1] }, then: 1 },
-                { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 2] }, then: 4 },
-                { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 3] }, then: 7 },
-              ],
-              default: 10,
-            },
-          },
-          day: 1,
-        },
-      };
-    } else if (d === 'firstDayOfIsoWeek') {
-      addFields[newColumns[i]] = {
-        // We subtract to the target date a number of days corresponding to (isoDayOfWeek - 1)
-        $subtract: [
-          $$(step.column),
-          {
-            $multiply: [
-              { $subtract: [{ $isoDayOfWeek: $$(step.column) }, 1] },
-              24 * 60 * 60 * 1000,
-            ],
-          },
-        ],
-      };
-    } else if (d === 'previousDay') {
-      // We subtract to the target date 1 day in milliseconds
-      addFields[newColumns[i]] = { $subtract: [$$(step.column), 24 * 60 * 60 * 1000] };
-    } else if (d === 'firstDayOfPreviousYear') {
-      addFields[newColumns[i]] = {
-        $dateFromParts: {
-          year: { $subtract: [{ $year: $$(step.column) }, 1] },
-          month: 1,
-          day: 1,
-        },
-      };
-    } else if (d === 'firstDayOfPreviousMonth') {
-      addFields[newColumns[i]] = {
-        $dateFromParts: {
-          year: {
-            $cond: [
-              { $eq: [{ $month: $$(step.column) }, 1] },
-              { $subtract: [{ $year: $$(step.column) }, 1] },
-              { $year: $$(step.column) },
-            ],
-          },
-          month: {
-            $cond: [
-              { $eq: [{ $month: $$(step.column) }, 1] },
-              12,
-              { $subtract: [{ $month: $$(step.column) }, 1] },
-            ],
-          },
-          day: 1,
-        },
-      };
-    } else if (d === 'firstDayOfPreviousWeek') {
-      addFields[newColumns[i]] = {
-        // We subtract to the target date a number of days corresponding to (dayOfWeek - 1)
-        $subtract: [
-          { $subtract: [$$(step.column), 7 * 24 * 60 * 60 * 1000] },
-          {
-            $multiply: [{ $subtract: [{ $dayOfWeek: $$(step.column) }, 1] }, 24 * 60 * 60 * 1000],
-          },
-        ],
-      };
-    } else if (d === 'firstDayOfPreviousQuarter') {
-      addFields[newColumns[i]] = {
-        $dateFromParts: {
-          year: {
-            $cond: [
-              { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 1] },
-              { $subtract: [{ $year: $$(step.column) }, 1] },
-              { $year: $$(step.column) },
-            ],
-          },
-          month: {
-            $switch: {
-              branches: [
-                { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 1] }, then: 10 },
-                { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 2] }, then: 1 },
-                { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 3] }, then: 4 },
-              ],
-              default: 7,
-            },
-          },
-          day: 1,
-        },
-      };
-    } else if (d === 'firstDayOfPreviousIsoWeek') {
-      addFields[newColumns[i]] = {
-        // We subtract to the target date a number of days corresponding to (isoDayOfWeek - 1)
-        $subtract: [
-          { $subtract: [$$(step.column), 7 * 24 * 60 * 60 * 1000] },
-          {
-            $multiply: [
-              { $subtract: [{ $isoDayOfWeek: $$(step.column) }, 1] },
-              24 * 60 * 60 * 1000,
-            ],
-          },
-        ],
-      };
-    } else if (d === 'previousYear') {
-      addFields[newColumns[i]] = { $subtract: [{ $year: $$(step.column) }, 1] };
-    } else if (d === 'previousMonth') {
-      addFields[newColumns[i]] = {
-        $cond: [
-          { $eq: [{ $month: $$(step.column) }, 1] },
-          12,
-          { $subtract: [{ $month: $$(step.column) }, 1] },
-        ],
-      };
-    } else if (d === 'previousWeek') {
-      // We subtract to the target date 7 days in milliseconds
-      addFields[newColumns[i]] = {
-        $week: { $subtract: [$$(step.column), 7 * 24 * 60 * 60 * 1000] },
-      };
-    } else if (d === 'previousQuarter') {
-      addFields[newColumns[i]] = {
-        $switch: {
-          branches: [
-            { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 1] }, then: 4 },
-            { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 2] }, then: 1 },
-            { case: { $lte: [{ $divide: [{ $month: $$(step.column) }, 3] }, 3] }, then: 2 },
-          ],
-          default: 3,
-        },
-      };
-    } else if (d === 'previousIsoWeek') {
-      // We subtract to the target date 7 days in milliseconds
-      addFields[newColumns[i]] = {
-        $isoWeek: { $subtract: [$$(step.column), 7 * 24 * 60 * 60 * 1000] },
-      };
+    if (d in ADVANCED_DATE_EXTRACT_MAP) {
+      addFields[newColumns[i]] = ADVANCED_DATE_EXTRACT_MAP[d as S.AdvancedDateInfo](step);
     } else {
-      //
       addFields[newColumns[i]] = {
         [`${DATE_EXTRACT_MAP[d as S.BasicDatePart]}`]: $$(step.column),
       };
